requirements.txt
Flask
flask-cors
requests
python-docx
PyMuPDF
faiss-cpu
sentence-transformers
gunicorn
